SentinEL/requirements.txt
fastapi
uvicorn
google-cloud-bigquery
google-cloud-aiplatform
pydantic
